[DEV-154] rename: loadNoteModal 컴포넌트 노트 디렉토리로 이동
